Allow key provision via query parameter
This is so that you can authenticate by providing
the key in the URL, which is necessary for the
Apple's Calendar because if you want to subscribe
to a calander, it only asks for a URL and doesn't
provide a way to authenticate with headers.

When using this method, the URL should then be
considered as private (can't be shared to others)
and the server should be behind an HTTPS protocol.

vin01 does point out that private URls may not
be so private, in his blog post from 7 May 2024
titled "You can not simply publicly access private
secure links, can you?" ¹

¹ https://vin01.github.io/piptagole/security-tools/soar/urlscan/hybrid-analysis/data-leaks/urlscan.io/cloudflare-radar%22/2024/03/07/url-database-leaks-private-urls.html