Use user_bp.url_prefix instead of hardcoding
Helps keep things DRY. I also found that in
hello.py I wasn't using the prefix - it was
routing to "/v1/<user>/hello" instead of
"/v1/user/<user>/hello".

This creates a circular dependency with user.py
and auth.py, but it doesn't seem to be a problem.

I'm keeping the user url prefix hardcoded in the
tests as it makes them more readable, in my
opinion.